Implemented layout asset in Unreal
This layout is made to work with Blender. The layout exports a json file with the reference to the blend files in ftrack that generated the fbx that has been imported in Unreal. In Blender, when the layout is loaded, pype will load the assets from the blend files and set the transform from the layout.