Removed `apply_settings` override to fix plugin settings actually being applied
This makes it so again that it's up to the studio admin to enable or disable the plug-in per project instead of whether the modeler's workfile just happened to have the current renderer set to arnold or not. Note that the plug-in is currently disabled by default in project settings.