Remove MRV2 backwards compatibility
With latest 0.8.2+ session related functions moved into a `mrv2.session` module. There's been so much refactoring in the MRV2 API that it'd become code spaghetti to maintain all whereas the features we actually want to use won't be available in older versions anyway. So we're much better of at least assuming the user will have 0.8.2+ as MRV2 version.