Fusion: provide better logging for validate saver crash due type error (#6082)
* OP-7467 - move get_tool_resolution to classmethod

Doesn't make much sense to have it outside of class. Debugging of it is impossible there (because of missing logger). Imho.

* OP-7467 - add check for frame

Limits uncaught error when resolution info is None, which could happen when saver is not connected.

* OP-7467 - remove debugging messages

* OP-7467 - enhance get_invalid

Handle exception to select broken saver in the Publisher UI.

* OP-7467 - refactor check

Here it makes more sense. We try to run some expression, but it might still result in None.

---------

Co-authored-by: Libor Batek <112623825+LiborBatek@users.noreply.github.com>
Co-authored-by: Jakub Ježek <jakubjezek001@gmail.com>